Removeesconversion helpers, addesunit test,  updates prediction calculation

+++ UTAT_Compression_Algorithm/C/compression.h
@@ -1,7 +1,14 @@
 #ifndef COMPRESSION
 #define COMPRESSION
 
-double predictionCalculation(gsl_vector* local_d, gsl_vector* weight, double local_sum, int t, double data, double* dr_sample_value, double* pred_sample_value);
-double localSum(int x, int y, int Nx, gsl_matrix* data);
+#include <gsl/gsl_matrix.h>
+#include <gsl/gsl_vector.h>
+#include <gsl/gsl_blas.h>
+#include <math.h>
+
+#include "helperlib.h"
+
+int predictionCalculation(gsl_vector* local_d, gsl_vector* weight, int local_sum, int t, int data, int* dr_sample_value, int* pred_sample_value);
+int localSum(int x, int y, int Nx, gsl_matrix* data);
 
 #endif
